Add Show Userlist to FavHub's Dialog (in Chat&Misc) Some copyright stuff Some Fixies

+++ dcpp/ADLSearch.h
@@ -56,8 +56,8 @@ public:
 	
 	// dest dir index
 	unsigned long ddIndex;
-	int adlsRaw;
-	int adlsPoints;
+	int64_t adlsRaw;
+	int64_t adlsPoints;
 	/// Search source type
 	enum SourceType {
 		TypeFirst = 0,
